+++ default.go
@@ -25,14 +25,14 @@ var (
 
 	// DefaultMaskFormatter defines the mask formatter function used to format
 	// mask message when the function is not defined
-	DefaultMaskFormatter = func(erw ErrorWrapper) string {
+	DefaultMaskFormatter MaskFormatter = func(erw ErrorWrapper) string {
 		return erw.RawMaskMessage()
 	}
 
 	// DefaultMessageFormatter defines the formatter function used to format
 	// formatted plain or mask message when the function is not defined.
 	// In default, this function will set error code to plain or mask message.
-	DefaultMessageFormatter = func(msg string, e ErrorWrapper) string {
+	DefaultMessageFormatter MessageFormatter = func(msg string, e ErrorWrapper) string {
 		return fmt.Sprintf("%s (%d)", msg, e.Code())
 	}