+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.3] - 2021-11-18
+
+### Changed
+
+- Change `ErrorDefinition.maskMessage` type to `*string` 
+- Change `ErrorDefinition.formatter` type to `*MessageFormatter` 
+- Change `ErrorDefinition.maskFormatter` type to `*MaskFormatter` 
+
 ## [0.0.2] - 2021-10-26
 
 ### Added
@@ -39,3 +47,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 [Unreleased]: https://github.com/rapidashorg/errwrap/compare/v0.0.1...HEAD
 [0.0.1]: https://github.com/rapidashorg/errwrap/releases/tag/v0.0.1
+[0.0.2]: https://github.com/rapidashorg/errwrap/releases/tag/v0.0.2
+[0.0.3]: https://github.com/rapidashorg/errwrap/releases/tag/v0.0.3